* commit '3f1974031c88750a14adc8f2f49538837238abf9':
  Support user mode HDMI timing setting

+++ common/planes/DisplayPlane.cpp
@@ -253,7 +253,7 @@ BufferMapper* DisplayPlane::mapBuffer(DataBuffer *buffer)
     BufferManager *bm = Hwcomposer::getInstance().getBufferManager();
 
     // don't map buffer if cache was full
-    if (mDataBuffers.size() >= mCacheCapacity) {
+    if ((int)mDataBuffers.size() >= mCacheCapacity) {
         return NULL;
     }
